# [IMP] website: reduce SQL queries to serve pages