--- a/addons/website/models/ir_http.py
+++ b/addons/website/models/ir_http.py
@@ -227,7 +227,10 @@ class Http(models.AbstractModel):
         # If the company of the website is not in the allowed companies of the user, set the main
         # company of the user.
         website_company_id = request.website._get_cached('company_id')
-        if website_company_id in request.env.user.company_ids.ids:
+        if request.website.is_public_user():
+            # avoid a read on res_company_user_rel in case of public user
+            context['allowed_company_ids'] = [website_company_id]
+        elif website_company_id in request.env.user.company_ids.ids:
             context['allowed_company_ids'] = [website_company_id]
         else:
             context['allowed_company_ids'] = request.env.user.company_id.ids