--- a/addons/website/models/website.py
+++ b/addons/website/models/website.py
@@ -1357,6 +1357,7 @@ class Website(models.Model):
             'user_id': self.user_id.id,
             'company_id': self.company_id.id,
             'default_lang_id': self.default_lang_id.id,
+            'homepage_id': self.homepage_id.id,
         }
 
     def _get_cached(self, field):