--- a/addons/website/models/website_menu.py
+++ b/addons/website/models/website_menu.py
@@ -96,7 +96,7 @@ class Menu(models.Model):
 
     def write(self, values):
         res = super().write(values)
-        if 'website_id' in values or 'group_ids' in values or 'sequence' in values:
+        if 'website_id' in values or 'group_ids' in values or 'sequence' in values or 'page_id' in values:
             self.clear_caches()
         return res