--- a/addons/website/tests/test_performance.py
+++ b/addons/website/tests/test_performance.py
@@ -89,31 +89,31 @@ class TestWebsitePerformance(UtilPerf):
 
     def test_10_perf_sql_queries_page(self):
         # standard untracked website.page
-        self.assertEqual(self._get_url_hot_query(self.page.url), 8)
-        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 11)
+        self.assertEqual(self._get_url_hot_query(self.page.url), 7)
+        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 10)
         self.menu.unlink()
-        self.assertEqual(self._get_url_hot_query(self.page.url), 10)
-        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 13)
+        self.assertEqual(self._get_url_hot_query(self.page.url), 9)
+        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 12)
 
     def test_15_perf_sql_queries_page(self):
         # standard tracked website.page
         self.page.track = True
-        self.assertEqual(self._get_url_hot_query(self.page.url), 16)
-        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 19)
+        self.assertEqual(self._get_url_hot_query(self.page.url), 15)
+        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 18)
         self.menu.unlink()
-        self.assertEqual(self._get_url_hot_query(self.page.url), 18)
-        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 21)
+        self.assertEqual(self._get_url_hot_query(self.page.url), 17)
+        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 20)
 
     def test_20_perf_sql_queries_homepage(self):
         # homepage "/" has its own controller
-        self.assertEqual(self._get_url_hot_query('/'), 17)
-        self.assertEqual(self._get_url_hot_query('/', cache=False), 18)
+        self.assertEqual(self._get_url_hot_query('/'), 16)
+        self.assertEqual(self._get_url_hot_query('/', cache=False), 17)
 
     def test_30_perf_sql_queries_page_no_layout(self):
         # website.page with no call to layout templates
         self.page.arch = '<div>I am a blank page</div>'
-        self.assertEqual(self._get_url_hot_query(self.page.url), 8)
-        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 9)
+        self.assertEqual(self._get_url_hot_query(self.page.url), 7)
+        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 8)
 
     def test_40_perf_sql_queries_page_multi_level_menu(self):
         # menu structure should not impact SQL requests
@@ -131,8 +131,8 @@ class TestWebsitePerformance(UtilPerf):
         menu_bb.parent_id = menu_b
         menu_aa.parent_id = menu_a
 
-        self.assertEqual(self._get_url_hot_query(self.page.url), 8)
-        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 11)
+        self.assertEqual(self._get_url_hot_query(self.page.url), 7)
+        self.assertEqual(self._get_url_hot_query(self.page.url, cache=False), 10)
 
     def test_50_perf_sql_web_assets(self):
         # assets route /web/assets/..