--- a/addons/website/tests/test_performance.py
+++ b/addons/website/tests/test_performance.py
@@ -106,7 +106,7 @@ class TestWebsitePerformance(UtilPerf):
 
     def test_20_perf_sql_queries_homepage(self):
         # homepage "/" has its own controller
-        self.assertEqual(self._get_url_hot_query('/'), 16)
+        self.assertEqual(self._get_url_hot_query('/'), 15)
         self.assertEqual(self._get_url_hot_query('/', cache=False), 17)
 
     def test_30_perf_sql_queries_page_no_layout(self):