Multiple one line improvements to reduce SQL Queries to serve pages.
Not only it does reduce SQL Queries but it mainly completely remove the read of some tables like lang, website, company_rel, translation etc.

Basically, this is removing ~4 queries in almost every serve page flows (regular page / homepage / empty page / page with no menu).
This is quite good as some of those flows are lowered from 10 requests to 6.