@rdeodoo 2022-02-23
@rco-odoo This is the PR including the commit I told you about, see `[[IMP] models.py, *: don't read ir_translation table if empty source` commit.

Could you approve/review that commit? Feel free to give your opinion about other commits if you want to.