@rdeodoo 2022-02-24
> Because we use request.uid in the orm-cache of _get_menu_page_ids we have to be extra-careful to not call that method prior of ir.http _authenticate() where request.uid may change (e.g. None->public in @route(..., auth='public')). In master this mean we cannot use the function in an ir.http _match() override.

Thanks for the heads up, I'll add a comment about that in master I guess

> The query-count will ensure we don't introduce non-_get_cached() values back. At least not in the modules covered by the tests. +1

Yep, I introduced that yesterday at https://github.com/odoo/odoo/commit/4795e325966a70dc8b3653caca69c00432aab93e :+1: 

> Do you want I test your branch?

Feel free to do it, can't hurt. Even if we now have a wider testing suite since https://github.com/odoo/odoo/commit/4795e325966a70dc8b3653caca69c00432aab93e and that those one-liner seems pretty safe.