@rdeodoo 2022-02-25
Hello @Yenthe666 , most likely not.
The risk/reward would not be worth it in those version.
Arguably, it was already on the edge in 15.0, as no one-liner is ever 100% safe, but as it is our last main stable version, having a whole "year-version" benefiting from it was ok, IMO.