@odony 2022-02-25
> The risk/reward would not be worth it in those version. 

Actually that's a good point. For this sort of improvement, it's always interesting to know the magnitude of the benefit that is expected. Outside of the raw count of queries, how much of a speedup were you able to measure, before/after, in terms of percentage of the time?