@rdeodoo 2022-02-25
Here is a benchmark using `ab`

Before
```

Non-tracked homepage [BEFORE]

Complete requests:      1000
Requests per second:    72.02 [#/sec] (mean)
Time per request:       13.884 [ms] (mean)

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       1
Processing:    11   14   4.1     12      49
Waiting:       10   14   4.1     12      49
Total:         11   14   4.1     12      49

Percentage of the requests served within a certain time (ms)
  50%     12
  66%     13
  75%     14
  80%     14
  90%     19
  95%     22
  98%     26
  99%     32
 100%     49 (longest request)
```

After
```
Complete requests:      1000
Requests per second:    99.35 [#/sec] (mean)
Time per request:       10.065 [ms] (mean)

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:     9   10   0.8     10      16
Waiting:        8   10   0.8     10      16
Total:          9   10   0.8     10      16

Percentage of the requests served within a certain time (ms)
  50%     10
  66%     10
  75%     11
  80%     11
  90%     11
  95%     11
  98%     12
  99%     12
 100%     16 (longest request)
```

So in average, from ~13ms to ~10ms. Won't be a game killer, a few ms are nothing when serving a page and we have surely more bigger bottleneck python wise to solve, especially when adding other website business modules on top of website.

But I think it is still something, especially to remove SQL Queries if we can do without it in low level methods. We came a long way since priori to https://github.com/odoo/odoo/pull/47257 (see PR desc for stats).
Following would be to check if/how we could improve the visitor tracking part in master.