@odony 2022-02-25
> Here is a benchmark using `ab`

Thanks! It's useful, in order to keep things in proportion :+1: 

As long as we don't increase code complexity, even small gains are good, as they might compensate for small losses elsewhere, and should add up.