@odony 2022-03-04
> @odony would you allow a backport?

Well, you've seen the numbers, it's more of a long-term work relevant in master than in any stable branch.

I suggest you test it: apply the patch on a real deployment, and compare the avg/p95 response times over a week, before vs after the patch. Keep in mind that a typical network latency is at least 15-20ms (up to 200-300ms overseas) and that TLS setup takes 2 roundtrips when you're opening a new connection. So a 3ms difference would be buried in the network jitter even in the best case scenarios, when taken on its own.